requirements.txt
tensorflow
numpy
streamlit
matplotlib
pillow
scipy
